Pause menu now works
