+++ Assets/Materials/Materials/Hand Material.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: be87337a9fbada140be44879b38b7576
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
