--- Assets/compassNeedle.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 48c0baf5d8d139047ba0555530cc086d
