+++ ProjectSettings/EditorBuildSettings.asset
@@ -5,7 +5,7 @@ EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Scenes:
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/mainMenu.unity
     guid: b57c05799619e054695b8f0cac5a3cfd
   - enabled: 1
